+++ packages/formatting/eslint-config-react-typescript/index.js
@@ -85,8 +85,20 @@ module.exports = {
     'import/order': [
       ERROR,
       {
+        alphabetize: {
+          order: 'asc',
+          caseInsensitive: false
+        },
         groups: ['builtin', 'external', 'internal', ['parent', 'sibling'], 'index'],
-        'newlines-between': 'never'
+        'newlines-between': 'always',
+        pathGroups: [
+          {
+            pattern: 'react',
+            group: 'external',
+            position: 'before'
+          }
+        ],
+        pathGroupsExcludedImportTypes: ['react']
       }
     ],